pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "stubhub"
dynamic = ["version"]
description = ""
readme = "README.md"
requires-python = ">=3.11"
# Don't need dependencies, those should be installed in the container
# and are managed by pipenv
dependencies = [
    "fake-useragent>=2.2.0",
    "httpx>=0.27.2",
    "prefect[aws]>=3.4.25",
    "prefecto>=1.1.1",
    "pydantic-mongo>=3.1.0",
    "pymongo>=4.15.3",
]
license.file = "LICENSE"

[project.scripts]
stubhub = "stubhub.cli:cli"

[tool.interrogate]
exclude = ["tests/"]
ignore-init-module = true

[tool.pytest.ini_options]
pythonpath = [
  "src"
]

[dependency-groups]
dev = [
    "click>=8.1.7",
    "pre-commit>=4.0.1",
    "pytest>=8.3.3",
]
